// src/models/ICounter.ts
export interface ICounter {
    all: number;
    newTrainers: number;
    newStudents: number;
    allStudents: number;
    students: number;
    trainers: number;
    admins: number;
    archive: number;
    programmers: number;
    newTrainersEdited: number,
    newStudentsEdited: number,
    studentsEdited: number,
    studentsRequisiteEdited: number,
    trainersEdited: number,
    adminsEdited: number,
    debtors: number,
    archiveStudents: number,
    testUsers: number,
    archivedTestUsers: number,
    directors: number,
    zDirectors: number
}